+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.29.0] - 2020-05-25
 ### Added
 - VIMEET-2117 - Webinar audio/video settings
 - VIMEET-2088 - Sort sheet participants